backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
mysqlclient==2.2.0
tensorflow==2.15.0
pillow==10.0.1
celery==5.3.4
redis==5.0.1
